Separated technology data classes for VIRL and WebEx into their own files.
