+++ features/VIRL_chat.js
@@ -4,447 +4,15 @@
  */
 const {BotkitConversation} = require('botkit');
 const request = require('request-promise');
+const WEBEX_methods = require('./WEBEX_methods_class.js');
+const WEBEX_data = require('./WEBEX_data_class.js');
+const VIRL_methods = require('./VIRL_methods_class.js');
+const VIRL_data = require('./VIRL_data_class.js');
 virlServers = [];
 virlServers = process.env.SERVER_LIST.split(',');
 const VIRL_USERNAME = process.env.VIRL_USERNAME;
 const VIRL_PASSWORD = process.env.VIRL_PASSWORD;
 
-function VIRL_data(username_virl, password_virl, server_name) {
-    this.username_virl = username_virl;
-    this.password_virl = password_virl;
-    this.server_name = server_name;
-}
-
-function VIRL_methods(obj) {
-    this.get_token = async function () {
-        let authenticate_options = {
-            method: 'POST',
-            uri: obj.server_name + '/api/v0/authenticate',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json'
-                },
-            body: {username: obj.username_virl, password: obj.password_virl},
-            json: true
-        };
-        let response = async () =>
-            request(authenticate_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return ''
-            });
-        obj.bearer = await response();
-    };
-    this.randomPasswordGenerator = function () {
-        let charList = "abcdefghijkmnpqrstuvwxyz0123456789ABCDEFGHJKLMNPQRSTUVWXYZ0123456789";
-        let listLength = charList.length;
-        let passwordString = '';
-        let i;
-        for (i = 0; i < 9; i++) {
-            let x = Math.floor((Math.random() * listLength));
-            passwordString += charList[x];
-        }
-        return passwordString
-    };
-    this.list_users = async () => {
-        let list_users_options = {
-            method: 'GET',
-            uri: obj.server_name + '/api/v0/users',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(list_users_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        let results = await response();
-        obj.users = [];
-        for (let [key, value] of Object.entries(results)) {
-            obj.users.push(results[key].username);
-        }
-        console.log(obj.users);
-    };
-    this.add_user = async (username_webex, user_email, new_password) => {
-        let add_user_options = {
-            method: 'POST',
-            uri: obj.server_name + '/api/v0/users/' + username_webex,
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            body: {
-                "password": new_password,
-                "fullname": user_email,
-                "description": "",
-                "roles": [
-                    "User"
-                    // "admin"
-                ],
-                "context": {}
-            },
-            json: true
-        };
-        let response = async () =>
-            request(add_user_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        obj.result = await response();
-    };
-    this.delete_user = async (username_webex) => {
-        let delete_user_options = {
-            method: 'DELETE',
-            uri: obj.server_name + '/api/v0/users/' + username_webex,
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(delete_user_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        obj.result = await response();
-    };
-    this.change_password = async (username_webex, new_password) => {
-        let change_password_options = {
-            method: 'PUT',
-            uri: obj.server_name + '/api/v0/users/' + username_webex + '/change_password',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            body: {
-                "old_password": '',
-                "new_password": new_password
-            },
-            json: true
-        };
-        let response = async () =>
-            request(change_password_options).then(function (resp) {
-                // A null response is a good response for this API call
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        obj.result = await response();
-    };
-    this.list_labs = async (username_webex) => {
-        let list_labs_options = {
-            method: 'GET',
-            uri: obj.server_name + '/api/v0/diagnostics',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(list_labs_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        let result = await response();
-
-        let labs = result.user_roles.labs_by_user[username_webex];
-        console.log(labs);
-        if (labs) {
-            let epoch_time_now = Math.floor(new Date().getTime() / 1000.0);
-            let results_list = [];
-            let x;
-            let labs_str = '';
-            for (x = 0; x < labs.length; x++) {
-                let created_seconds = Math.floor(result.labs[labs[x]].created);
-                let seconds = epoch_time_now - created_seconds;
-                let days = Math.floor(seconds / (3600 * 24));
-                seconds -= days * 3600 * 24;
-                let hrs = Math.floor(seconds / 3600);
-                seconds -= hrs * 3600;
-                let mnts = Math.floor(seconds / 60);
-                seconds -= mnts * 60;
-                let uptime = days + " days, " + hrs + " Hrs, " + mnts + " Minutes, " + seconds + " Seconds";
-
-                labs_str = '        **  Lab id: ' + labs[x] + '        Uptime: ' + uptime + '\n';
-                results_list.push(labs_str);
-            }
-            obj.list_labs = results_list;
-        } else {
-            obj.list_labs = []
-        }
-    };
-    this.list_labs_all = async () => {
-        let list_labs_all_options = {
-            method: 'GET',
-            uri: obj.server_name + '/api/v0/diagnostics',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(list_labs_all_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        let result = await response();
-        let epoch_time_now = Math.floor(new Date().getTime() / 1000.0);
-        let results_list = [];
-        let labs_str = '';
-        let labs;
-        for (let [key, value] of Object.entries(result.user_roles.labs_by_user)) {
-            labs = result.user_roles.labs_by_user[key];
-            if (labs.length > 0) {
-                results_list.push('    ' + '* Labs for account: ' + key + '\n');
-                let x;
-                for (x = 0; x < labs.length; x++) {
-                    let created_seconds = Math.floor(result.labs[labs[x]].created);
-                    let seconds = epoch_time_now - created_seconds;
-                    let days = Math.floor(seconds / (3600 * 24));
-                    seconds -= days * 3600 * 24;
-                    let hrs = Math.floor(seconds / 3600);
-                    seconds -= hrs * 3600;
-                    let mnts = Math.floor(seconds / 60);
-                    seconds -= mnts * 60;
-                    let uptime = days + " days, " + hrs + " Hrs, " + mnts + " Minutes, " + seconds + " Seconds";
-
-                    labs_str = '            **  Lab id: ' + labs[x] + '        Uptime: ' + uptime + '\n';
-                    results_list.push(labs_str);
-                }
-            }
-        }
-        obj.list_labs_all = results_list;
-    };
-    this.list_user_labs = async () => {
-        let list_users_options = {
-            method: 'GET',
-            uri: obj.server_name + '/api/v0/labs',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(list_users_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        obj.labs = await response();
-    };
-    this.get_lab_details = async (lab_id) => {
-        let get_lab_details_options = {
-            method: 'GET',
-            uri: obj.server_name + '/api/v0/labs/' + lab_id,
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(get_lab_details_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-
-        obj.lab_details.push(await response());
-    };
-    this.delete_lab = async function (lab_id) {
-        let delete_lab_options = {
-            method: 'DELETE',
-            uri: obj.server_name + '/api/v0/labs/' + lab_id,
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(delete_lab_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        obj.result = await response();
-    };
-    this.list_labs_ids = async (username_webex) => {
-        let list_labs_ids_options = {
-            method: 'GET',
-            uri: obj.server_name + '/api/v0/diagnostics',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(list_labs_ids_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        let result = await response();
-        labs = result.user_roles.labs_by_user[username_webex];
-        if (labs) {
-            obj.list_labs_ids = labs;
-        } else {
-            obj.list_labs_ids = []
-        }
-    };
-    this.get_system_status = async () => {
-        let get_system_status_options = {
-            method: 'GET',
-            uri: obj.server_name + '/api/v0/system_stats',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    'Accept': 'application/json',
-                    'cache-control': "no-cache",
-                    "Authorization": "Bearer " + obj.bearer
-                },
-            json: true
-        };
-        let response = async () =>
-            request(get_system_status_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        obj.system_status = await response();
-    };
-}
-
-function WEBEX_data() {
-    this.webex_token = process.env.ACCESS_TOKEN
-}
-
-function WEBEX_methods(obj) {
-    this.get_webex_email_and_username = async function () {
-        let authenticate_options = {
-            method: 'GET',
-            uri: 'https://api.ciscospark.com/v1/people?id=' + obj.webex_user_id,
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    "Authorization": "Bearer " + obj.webex_token
-                },
-            json: true
-        };
-        let response = async () =>
-            request(authenticate_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        let result = await response();
-        obj.webex_user_email = result.items[0].emails[0];
-        obj.webex_username = obj.webex_user_email.substring(0, obj.webex_user_email.lastIndexOf("@"));
-        obj.display_name = result.items[0].displayName;
-    };
-    this.get_id_from_email = async function (email) {
-        let web_email = email.replace(/@/, "%40");
-        let get_id_from_email_options = {
-            method: 'GET',
-            uri: 'https://api.ciscospark.com/v1/people?email=' + web_email,
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    "Authorization": "Bearer " + obj.webex_token
-                },
-            json: true
-        };
-        let response = async () =>
-            request(get_id_from_email_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        let result = await response();
-        obj.webex_user_id = result.items[0].id;
-    };
-    this.send_message = async function (message) {
-        let message_options = {
-            method: 'POST',
-            uri: 'https://api.ciscospark.com/v1/messages',
-            rejectUnauthorized: false,
-            headers:
-                {
-                    'Content-Type': 'application/json',
-                    "Authorization": "Bearer " + obj.webex_token
-                },
-            body: {
-                "toPersonId": obj.webex_user_id,
-                "text": message,
-            },
-            json: true
-        };
-        let response = async () =>
-            request(message_options).then(function (resp) {
-                return resp
-            }).catch(function (err) {
-                return err
-            });
-        let result = await response();
-    };
-}
 
 module.exports = function (controller) {
     /*
@@ -725,7 +293,6 @@ module.exports = function (controller) {
             await virl_methods.get_token();
 
             if (virl_data.bearer === '') {
-                // convo.gotoThread('bad_password')
                 lab_string += '\nLogin failed for: ' + virlServers[n]
             } else {
                 virl_data.lab_details = [];
@@ -784,7 +351,6 @@ module.exports = function (controller) {
         }
     }, {key: 'virl_server'});
 
-    // delete_lab_dialog.say('{{vars.labs}}');
     delete_lab_dialog.ask("Please enter the Lab ID of the lab you would like to delete or 'quit' to exit: ", async (response, convo, bot) => {
         if (convo.vars.lab_id === 'quit') {
             convo.gotoThread('goodbye')
@@ -850,7 +416,6 @@ module.exports = function (controller) {
             await virl_methods.get_token();
 
             if (virl_data.bearer === '') {
-                // convo.gotoThread('bad_password')
                 lab_string += '\nLogin failed for: ' + virlServers[n]
             } else {
                 virl_data.lab_details = [];
@@ -879,7 +444,6 @@ module.exports = function (controller) {
     }, {key: 'virl_password'});
 
     details_lab_dialog.say('{{vars.labs}}');
-    // details_lab_dialog.addMessage('Your password was not accepted by server {{vars.virl_server}}', 'bad_password');
 
     controller.addDialog(details_lab_dialog);
     controller.hears('VIRL list my lab details', 'message,direct_message', async (bot, message) => {
