+++ features/VIRL_data_class.js
@@ -1,0 +1,7 @@
+function VIRL_data_class(username_virl, password_virl, server_name) {
+    this.username_virl = username_virl;
+    this.password_virl = password_virl;
+    this.server_name = server_name;
+}
+
+module.exports = VIRL_data_class;
