+++ features/webex_data_class.js
@@ -1,0 +1,5 @@
+function WEBEX_data_class() {
+    this.webex_token = process.env.ACCESS_TOKEN
+}
+
+module.exports = WEBEX_data_class;
